# backend/pkg/api/user.swagger.json
{
  "swagger": "2.0",
  "info": {
    "title": "user.proto",
    "version": "version not set"
  },
  "tags": [
    {
      "name": "UserService"
    }
  ],
  "host": "localhost:8080",
  "basePath": "/api",
  "schemes": [
    "http"
  ],
  "consumes": [
    "application/json"
  ],
  "produces": [
    "application/json"
  ],
  "paths": {
    "/api/user/edit": {
      "post": {
        "operationId": "UserService_Edit",
        "responses": {
          "200": {
            "description": "A successful response.",
            "schema": {
              "$ref": "#/definitions/apiEditResponse"
            }
          },
          "default": {
            "description": "An unexpected error response.",
            "schema": {
              "$ref": "#/definitions/rpcStatus"
            }
          }
        },
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/apiEditRequest"
            }
          }
        ],
        "tags": [
          "UserService"
        ]
      }
    },
    "/api/user/fill": {
      "post": {
        "operationId": "UserService_Fill",
        "responses": {
          "200": {
            "description": "A successful response.",
            "schema": {
              "$ref": "#/definitions/apiFillResponse"
            }
          },
          "default": {
            "description": "An unexpected error response.",
            "schema": {
              "$ref": "#/definitions/rpcStatus"
            }
          }
        },
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/apiFillRequest"
            }
          }
        ],
        "tags": [
          "UserService"
        ]
      }
    },
    "/api/user/login": {
      "post": {
        "operationId": "UserService_Login",
        "responses": {
          "200": {
            "description": "A successful response.",
            "schema": {
              "$ref": "#/definitions/apiLoginResponse"
            }
          },
          "default": {
            "description": "An unexpected error response.",
            "schema": {
              "$ref": "#/definitions/rpcStatus"
            }
          }
        },
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/apiLoginRequest"
            }
          }
        ],
        "tags": [
          "UserService"
        ]
      }
    },
    "/api/user/refresh": {
      "post": {
        "operationId": "UserService_Refresh",
        "responses": {
          "200": {
            "description": "A successful response.",
            "schema": {
              "$ref": "#/definitions/apiRefreshResponse"
            }
          },
          "default": {
            "description": "An unexpected error response.",
            "schema": {
              "$ref": "#/definitions/rpcStatus"
            }
          }
        },
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/apiRefreshRequest"
            }
          }
        ],
        "tags": [
          "UserService"
        ]
      }
    },
    "/api/user/signup": {
      "post": {
        "operationId": "UserService_SignUp",
        "responses": {
          "200": {
            "description": "A successful response.",
            "schema": {
              "$ref": "#/definitions/apiSignUpResponse"
            }
          },
          "default": {
            "description": "An unexpected error response.",
            "schema": {
              "$ref": "#/definitions/rpcStatus"
            }
          }
        },
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/apiSignUpRequest"
            }
          }
        ],
        "tags": [
          "UserService"
        ]
      }
    }
  },
  "definitions": {
    "apiEditRequest": {
      "type": "object",
      "properties": {
        "ege": {
          "type": "integer",
          "format": "int32"
        },
        "gpa": {
          "type": "integer",
          "format": "int32"
        },
        "speciality": {
          "type": "string"
        },
        "eduType": {
          "type": "string"
        },
        "town": {
          "type": "string"
        },
        "financing": {
          "type": "string"
        }
      }
    },
    "apiEditResponse": {
      "type": "object"
    },
    "apiFillRequest": {
      "type": "object",
      "properties": {
        "ege": {
          "type": "integer",
          "format": "int32"
        },
        "gpa": {
          "type": "integer",
          "format": "int32"
        },
        "speciality": {
          "type": "string"
        },
        "eduType": {
          "type": "string"
        },
        "town": {
          "type": "string"
        },
        "financing": {
          "type": "string"
        }
      }
    },
    "apiFillResponse": {
      "type": "object"
    },
    "apiLoginRequest": {
      "type": "object",
      "properties": {
        "login": {
          "type": "string"
        },
        "password": {
          "type": "string"
        }
      }
    },
    "apiLoginResponse": {
      "type": "object",
      "properties": {
        "access": {
          "type": "string"
        },
        "refresh": {
          "type": "string"
        },
        "expiresIn": {
          "type": "string",
          "format": "int64"
        }
      }
    },
    "apiRefreshRequest": {
      "type": "object",
      "properties": {
        "refresh": {
          "type": "string"
        }
      }
    },
    "apiRefreshResponse": {
      "type": "object",
      "properties": {
        "access": {
          "type": "string"
        }
      }
    },
    "apiSignUpRequest": {
      "type": "object",
      "properties": {
        "login": {
          "type": "string"
        },
        "password": {
          "type": "string"
        }
      }
    },
    "apiSignUpResponse": {
      "type": "object"
    },
    "protobufAny": {
      "type": "object",
      "properties": {
        "@type": {
          "type": "string"
        }
      },
      "additionalProperties": {}
    },
    "rpcStatus": {
      "type": "object",
      "properties": {
        "code": {
          "type": "integer",
          "format": "int32"
        },
        "message": {
          "type": "string"
        },
        "details": {
          "type": "array",
          "items": {
            "type": "object",
            "$ref": "#/definitions/protobufAny"
          }
        }
      }
    }
  }
}